Refactor fixture matrix image comparison helpers
