+++ lib/fixture-matrix/steps/editor-validation-step.mjs
@@ -10,10 +10,7 @@ import {
   EDITOR_VALIDATE_BLOCKS_COMMAND,
   DEFAULT_EDITOR_VALIDATION_TARGET,
 } from '../shared/constants.mjs';
-
-function present(value) {
-  return value !== undefined && value !== null && String(value).trim() !== '';
-}
+import { firstPresent, fixtureStepMetadata } from './shared.mjs';
 
 export function editorBlockValidationStep(input = {}) {
   const fixture = input.fixture || {};
@@ -46,22 +43,10 @@ export function editorBlockValidationStep(input = {}) {
     command: EDITOR_VALIDATE_BLOCKS_COMMAND,
     allowFailure: true,
     args,
-    metadata: {
-      fixture_id: fixture.id,
-      fixture_path: fixture.fixture_path || fixture.directory,
-      phase: 'editor',
+    metadata: fixtureStepMetadata(fixture, 'editor', {
       ...(postId !== undefined ? { post_id: postId } : {}),
       ...(url !== undefined ? { url } : {}),
       ...(target !== undefined ? { target } : { target: DEFAULT_EDITOR_VALIDATION_TARGET }),
-    },
+    }),
   };
-}
-
-function firstPresent(values) {
-  for (const value of values) {
-    if (present(value)) {
-      return value;
-    }
-  }
-  return undefined;
 }
