+++ lib/fixture-matrix/steps/live-wp-parity-step.mjs
@@ -25,6 +25,7 @@ import {
   DEFAULT_VISUAL_PARITY_WAIT_FOR,
 } from '../shared/constants.mjs';
 import { isTruthySignal } from '../shared/utils.mjs';
+import { fixtureStepMetadata, resolveFixtureCandidateUrl } from './shared.mjs';
 
 // Compose `wordpress.capture-html` for a fixture's imported candidate front page.
 // The candidate URL resolution mirrors the visual-parity step: it defaults to `/`,
@@ -34,11 +35,7 @@ import { isTruthySignal } from '../shared/utils.mjs';
 export function liveWpParityCaptureStep(input = {}) {
   const fixture = input.fixture || {};
   const options = normalizeLiveWpParityRecipeOptions(input);
-  const candidateUrl = input.candidateUrl
-    || input.candidate_url
-    || fixture.candidate_url
-    || fixture.candidateUrl
-    || options.candidateUrl;
+  const candidateUrl = resolveFixtureCandidateUrl(input, fixture, options.candidateUrl);
   return {
     command: 'wordpress.capture-html',
     allowFailure: true,
@@ -50,12 +47,9 @@ export function liveWpParityCaptureStep(input = {}) {
       'network-policy=block',
       `wait-for=${options.waitFor}`,
     ],
-    metadata: {
-      fixture_id: fixture.id,
-      fixture_path: fixture.fixture_path || fixture.directory,
-      phase: 'capture-html',
+    metadata: fixtureStepMetadata(fixture, 'capture-html', {
       candidate_url: candidateUrl,
-    },
+    }),
   };
 }
