+++ lib/fixture-matrix/steps/recipe-builder.mjs
@@ -30,6 +30,7 @@ import { createFixtureMatrix, normalizeFixture, collectFixtureFiles } from '../f
 import { editorBlockValidationStep } from './editor-validation-step.mjs';
 import { visualParityCompareStep, normalizeVisualParityRecipeOptions } from './visual-parity-step.mjs';
 import { liveWpParityCaptureStep, liveWpParityEnabled } from './live-wp-parity-step.mjs';
+import { fixtureStepMetadata } from './shared.mjs';
 
 const CAPABILITY_PLUGIN_PROVISIONING = {
   'commerce-products': [
@@ -265,15 +266,6 @@ function fixtureCapabilityPlugins(fixture) {
     }
   }
   return plugins;
-}
-
-function fixtureStepMetadata(fixture, phase, extra = {}) {
-  return {
-    fixture_id: fixture.id,
-    fixture_path: fixture.fixture_path || fixture.directory,
-    phase,
-    ...extra,
-  };
 }
 
 function buildDependencyOverrideSetup(input, importer) {
