+++ lib/fixture-matrix/steps/visual-parity-step.mjs
@@ -18,6 +18,7 @@ import {
   VISUAL_PARITY_SOURCE_SUBDIR,
 } from '../shared/constants.mjs';
 import { objectValue, finiteNumber } from '../shared/utils.mjs';
+import { fixtureStepMetadata, resolveFixtureCandidateUrl } from './shared.mjs';
 
 // Compose the existing `wordpress.visual-compare` recipe command into a
 // per-fixture visual-parity step. This is the same command the reusable
@@ -52,11 +53,7 @@ export function visualParityCompareStep(input = {}) {
     || fixture.source_url
     || fixture.sourceUrl
     || `${options.sourceBaseUrl.replace(/\/+$/, '')}/${fixture.id || 'fixture'}/${sourceEntry}`;
-  const candidateUrl = input.candidateUrl
-    || input.candidate_url
-    || fixture.candidate_url
-    || fixture.candidateUrl
-    || options.candidateUrl;
+  const candidateUrl = resolveFixtureCandidateUrl(input, fixture, options.candidateUrl);
   const matrix = {
     comparisons: [
       {
@@ -80,13 +77,10 @@ export function visualParityCompareStep(input = {}) {
     args: [
       `matrix-json=${JSON.stringify(matrix)}`,
     ],
-    metadata: {
-      fixture_id: fixture.id,
-      fixture_path: fixture.fixture_path || fixture.directory,
-      phase: 'visual',
+    metadata: fixtureStepMetadata(fixture, 'visual', {
       source_url: sourceUrl,
       candidate_url: candidateUrl,
-    },
+    }),
   };
 }
